Zobrist hashing now produces uint instead of int

+++ scripts/core/pieces/Piece.cs
@@ -26,12 +26,12 @@ public class Piece(byte id, BasePiece basePiece, bool color, Vector2Int position
         return new Piece(Id, BasePiece, Color, Position, Movement, SpecialPieceType == SpecialPieceTypes.EN_PASSANTABLE_PAWN ? SpecialPieceTypes.PAWN : SpecialPieceType);
     }
 
-    public int GetZobristHash()
+    public uint GetZobristHash()
     {
         // Zobrist hash for piece is the zobrist hash for every IMovement it has XORed
         // XOR BasePiece hash
         // XOR SpecialPieceTypes hash
-        int result = 0;
+        uint result = 0;
         foreach (IMovement movement in Movement)
         {
             result ^= movement.GetZobristHash(color, Position);
@@ -44,7 +44,8 @@ public class Piece(byte id, BasePiece basePiece, bool color, Vector2Int position
 
     public override int GetHashCode()
     {
-        return GetZobristHash();
+        // Possible overflow? Should I worry abt this?
+        return (int)GetZobristHash();
     }
 
     private static readonly char[] files = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'];
